+++ lib/src/common.dart
@@ -22,11 +22,9 @@ final Canister ii          = Canister(Principal('rdmx6-jaaaa-aaaaa-aaadq-cai'));
 
 
 
-Future<double> check_icp_balance(String icp_id) async {
-    //Record record = Record.oftheMap({'account': Text(icp_id)});
-    //Uint8List sponse_bytes = await ledger.call(calltype: CallType.call, method_name: 'account_balance_dfx', put_bytes: c_forwards([record]));
+Future<IcpTokens> check_icp_balance(String icp_id, {CallType? calltype}) async {
     Uint8List sponse_bytes = await ledger.call(
-        calltype: CallType.call,
+        calltype: calltype ?? CallType.call,
         method_name: 'account_balance',
         put_bytes: c_forwards([
             Record.oftheMap({
@@ -34,30 +32,28 @@ Future<double> check_icp_balance(String icp_id) async {
             })
         ])
     );
-    Record icpts_balance_record = c_backwards(sponse_bytes)[0] as Record;
-    Nat64 e8s = icpts_balance_record['e8s'] as Nat64;
-    return e8s.value / BigInt.from(100000000); 
+    return IcpTokens.oftheRecord(c_backwards(sponse_bytes)[0] as Record);
 }
 
 
-Future<Nat64> transfer_icp(Caller caller, String fortheicpid, double mount, {double? fee, Nat64? memo, List<int>? subaccount_bytes, List<Legation> legations = const [] } ) async {
-    fee ??= 0.0001;
+Future<Variant> transfer_icp(Caller caller, String fortheicpid, IcpTokens mount, {IcpTokens? fee, Nat64? memo, List<int>? subaccount_bytes, List<Legation> legations = const [] } ) async {
+    fee ??= IcpTokens.oftheDoubleString('0.0001');
     memo ??= Nat64(BigInt.from(0));
     subaccount_bytes ??= Uint8List(32);
     if (subaccount_bytes.length != 32) { throw Exception(': subaccount_bytes-parameter of this function is with the length-quirement: 32-bytes.'); }
-    if (check_double_decimal_point_places(mount) > 8 || check_double_decimal_point_places(fee) > 8) {
-        throw Exception('mount and fee can have max: 8 decimal-point-number-places');
-    }
+
     Record sendargs = Record.oftheMap({
+        'to' : Blob(hexstringasthebytes(fortheicpid)),        
         'memo': memo,
-        'amount': Record.oftheMap({'e8s': Nat64(BigInt.from((mount * 100000000).toInt()))}),
-        'fee': Record.oftheMap({'e8s': Nat64(BigInt.from((fee * 100000000).toInt()))}),
-        'to': Text(fortheicpid),
-        'from_subaccount': Option(value: Blob(subaccount_bytes)),
-        // 'created_at_time': Option()
+        'amount': mount,
+        'fee': fee,
+        'from_subaccount': Option<Blob>(value: Blob(subaccount_bytes)),
+        'created_at_time': Option<Record>(value: Record.oftheMap({
+            'timestamp_nanos' : Nat64(get_current_time_nanoseconds())
+        }))
     });
-    Nat64 block_height = c_backwards(await ledger.call(calltype: CallType.call, method_name: 'send_dfx', put_bytes: c_forwards([sendargs]), caller: caller, legations: legations))[0] as Nat64;
-    return block_height;
+    Variant transfer_result = c_backwards(await ledger.call(calltype: CallType.call, method_name: 'transfer', put_bytes: c_forwards([sendargs]), caller: caller, legations: legations))[0] as Variant;
+    return transfer_result;
 }
 
 
@@ -80,26 +76,6 @@ String icp_id(Principal principal, {List<int>? subaccount_bytes}) {
     return text_format;
 }
 
-extension PrincipalIcpId on Principal {
-    String icp_id({List<int>? subaccount_bytes}) {
-        subaccount_bytes ??= Uint8List(32);
-        if (subaccount_bytes.length != 32) { throw Exception(': subaccount_bytes-parameter of this function is with the length-quirement: 32-bytes.'); }
-        List<int> blobl = [];
-        blobl.addAll(utf8.encode('\x0Aaccount-id'));
-        blobl.addAll(this.bytes);
-        blobl.addAll(subaccount_bytes);
-        Uint8List blob = Uint8List.fromList(sha224.convert(blobl).bytes);
-        Crc32 crc32 = Crc32();
-        crc32.add(blob);
-        List<int> text_format_bytes = [];
-        text_format_bytes.addAll(crc32.close());
-        text_format_bytes.addAll(blob);
-        String text_format = bytesasahexstring(text_format_bytes);
-        return text_format;
-    }
-
-}
-
 
 
 
@@ -120,83 +96,135 @@ Principal icpsubaccountbytes_as_a_principal(Uint8List subaccount_bytes) {
     return Principal.oftheBytes(Uint8List.fromList(principal_bytes));
 }
 
+Map<String, Never Function(CandidType)> transfer_error_match_map = {
+    'TxTooOld' : (allowed_window_nanos_r) {
+        throw Exception('TxTooOld');
+    },
+    'BadFee' : (expected_fee_r) {
+        throw Exception('BadFee, expected_fee: ${IcpTokens.oftheRecord((expected_fee_r as Record)['expected_fee'] as Record)}');
+    },
+    'TxDuplicate' : (duplicate_of_r) {
+        throw Exception('TxDuplicate, duplicate_of: ${((duplicate_of_r as Record)['duplicate_of'] as Nat64).value}');
+    },
+    'TxCreatedInFuture': (nul) {
+        throw Exception('TxCreatedInFuture');
+    },
+    'InsufficientFunds' : (balance_r) {
+        throw Exception('InsufficientFunds, balance: ${IcpTokens.oftheRecord((balance_r as Record)['balance'] as Record)}');
+    }
+};
 
-Future<Principal> create_canister(Caller caller, double icp_count, {Uint8List? from_subaccount_bytes, Nat64? block_height}) async {
+Map<String, Never Function(CandidType)> cmc_notify_error_match_map = {
+    'Refunded' : (refund_r) {
+        Record r = refund_r as Record;
+        throw Exception('Refunded, reason: ${(r['reason'] as Text).value}, block_index: ${r.find_option<Nat64>('block_index')}');
+    },
+    'InvalidTransaction' : (text) {
+        throw Exception('InvalidTransaction, ${(text as Text).value}');
+    },
+    'Other' : (rc) {
+        Record r = rc as Record;
+        throw Exception('error_message: ${(r['error_message'] as Text).value}, error_code: ${(r['error_code'] as Nat64).value}');
+    },
+    'Processing': (nul) {
+        throw Exception('Processing');
+    },
+    'TransactionTooOld' : (block) {
+        throw Exception('TransactionTooOld, block: ${(block as Nat64).value}');
+    }
+};
+
+
+Future<Principal> create_canister(Caller caller, IcpTokens icp_count, {Uint8List? from_subaccount_bytes, Nat64? block_height}) async {
     Uint8List to_subaccount_bytes = principal_as_an_icpsubaccountbytes(caller.principal);
     
     if (block_height == null) {
-        block_height = await transfer_icp(
+        block_height = match_variant<Nat64>(await transfer_icp(
             caller, 
-            cycles_mint.principal.icp_id(subaccount_bytes: to_subaccount_bytes), 
+            icp_id(cycles_mint.principal, subaccount_bytes: to_subaccount_bytes), 
             icp_count, 
             subaccount_bytes: from_subaccount_bytes,
             memo: MEMO_CREATE_CANISTER_nat64,
-        );
-        print('block_height: ${block_height}');
+        ), {
+            'Ok': (block_height) {
+                return block_height as Nat64;
+            },
+            'Err': (transfer_error) {
+                return match_variant<Never>(transfer_error as Variant, transfer_error_match_map);
+            }
+        });
+        print('block_height: ${block_height.value}');
     } else {
-        print('using given block_height: $block_height');
+        print('using given block_height: ${block_height.value}');
     }
 
-    Record notifycanisterargs = Record.oftheMap({
-        'block_height' : block_height,
-        'max_fee'      : Record.oftheMap({'e8s': Nat64(BigInt.from((0.0001 * 100000000).toInt()))}),
-        'to_canister'  : cycles_mint.principal.candid,
-        'to_subaccount': Option(value: Blob( to_subaccount_bytes )) 
+    Record notifycanisterarg = Record.oftheMap({
+        'controller' : caller.principal,
+        'block_index' : block_height,
     });
-    if (from_subaccount_bytes != null ) {
-        notifycanisterargs['from_subaccount'] = Option(value: Blob(from_subaccount_bytes));
-    }
 
-    Uint8List notify_sponse_bytes = await ledger.call(calltype: CallType.call, method_name: 'notify_dfx', put_bytes: c_forwards([notifycanisterargs]), caller: caller);
-    List<CandidType> notify_sponse_candids = c_backwards(notify_sponse_bytes);
-    Variant variant = notify_sponse_candids[0] as Variant;
-    if (variant.containsKey('CanisterCreated')) {
-        PrincipalReference principal_fer = variant['CanisterCreated'] as PrincipalReference;
-        Principal new_canister_principal = principal_fer.principal!;
-        print(new_canister_principal);
-        return new_canister_principal;
-    } else if (variant.containsKey('Refunded')) {
-        throw Exception('looks like the call was refunded for some reason. gave back this record: ${variant['Refunded']}');
-    } else {
-        throw Exception('notify_dfx gives-back this variant: ${variant}');
-    }
+    Variant notify_create_canister_result = c_backwards(await cycles_mint.call(
+        calltype: CallType.call, 
+        method_name: 'notify_create_canister', 
+        put_bytes: c_forwards([notifycanisterarg]), 
+        caller: caller
+    ))[0] as Variant;
+    
+    return match_variant<Principal>(notify_create_canister_result, {
+        'Ok': (p) {
+            return p as Principal;
+        },
+        'Err': (notify_error) {
+            return match_variant<Never>(notify_error as Variant, cmc_notify_error_match_map);
+        } 
+    });
 }
 
 
-Future<void> top_up_canister(Caller caller, double icp_mount, Principal canister_id, {Uint8List? from_subaccount_bytes, Nat64? block_height}) async {
+Future<Nat> top_up_canister(Caller caller, IcpTokens icp_mount, Principal canister_id, {Uint8List? from_subaccount_bytes, Nat64? block_height}) async {
     Uint8List to_subaccount_bytes = principal_as_an_icpsubaccountbytes(canister_id);
 
     if (block_height == null) {
-        block_height = await transfer_icp(
+        block_height = match_variant<Nat64>(await transfer_icp(
             caller, 
-            cycles_mint.principal.icp_id(subaccount_bytes: to_subaccount_bytes), 
+            icp_id(cycles_mint.principal, subaccount_bytes: to_subaccount_bytes), 
             icp_mount, 
             subaccount_bytes: from_subaccount_bytes,
             memo: MEMO_TOP_UP_CANISTER_nat64,
-        );
-        print('block_height: ${block_height}');
+        ), {
+            'Ok': (block_height) {
+                return block_height as Nat64;
+            },
+            'Err': (transfer_error) {
+                return match_variant<Never>(transfer_error as Variant, transfer_error_match_map);
+            }
+        });
+        print('block_height: ${block_height.value}');
     } else {
-        print('using given block_height: $block_height');
+        print('using given block_height: ${block_height.value}');
     }
 
-    Record notifycanisterargs = Record.oftheMap({
-        'block_height' : block_height,
-        'max_fee'      : Record.oftheMap({'e8s': Nat64(BigInt.from((0.0001 * 100000000).toInt()))}),
-        'to_canister'  : cycles_mint.principal.candid, 
-        'to_subaccount': Option(value: Blob( to_subaccount_bytes )) 
+    Record notifytopupargs = Record.oftheMap({
+        'block_index' : block_height,
+        'canister_id' : canister_id,
     });
-    if (from_subaccount_bytes != null ) {
-        notifycanisterargs['from_subaccount'] = Option(value: Blob(from_subaccount_bytes));
-    }
 
-    Variant variant = c_backwards(await ledger.call(calltype: CallType.call, method_name: 'notify_dfx', put_bytes: c_forwards([notifycanisterargs]), caller: caller))[0] as Variant;
-    if (variant.containsKey('ToppedUp')) {
-        return;
-    } else if (variant.containsKey('Refunded')) {
-        throw Exception('looks like the call was refunded for some reason. gave back this record: ${variant['Refunded']}');
-    } else {
-        throw Exception('notify_dfx gives-back this variant: ${variant}');
-    }
+    Variant notify_top_up_result = c_backwards(await cycles_mint.call(
+        calltype: CallType.call, 
+        method_name: 'notify_top_up', 
+        put_bytes: c_forwards([notifytopupargs]), 
+        caller: caller
+    ))[0] as Variant;
+    
+    return match_variant<Nat>(notify_top_up_result, {
+        'Ok': (nat) {
+            return nat as Nat;
+        },
+        'Err': (notify_error) {
+            return match_variant<Never>(notify_error as Variant, cmc_notify_error_match_map);
+        }
+    });
+
 }
 
 
@@ -265,8 +293,21 @@ class IcpTokens extends Record {
         while (s.length < IcpTokens.DECIMAL_PLACES + 1) { s = '0$s'; }
         int split_i = s.length - IcpTokens.DECIMAL_PLACES;
         s = '${s.substring(0, split_i)}.${s.substring(split_i)}';
-        while (s[s.length - 1] == '0' && s.length > 3/*minimum '0.0'*/) { s = s.substring(0, s.length - 1); }
+        while (s[s.length - 1] == '0' && s[s.length - 2] != '.') { s = s.substring(0, s.length - 1); }
         return s;   
+    }
+    IcpTokens round_decimal_places(int round_decimal_places) {
+        //print('icp without round: $this');
+        List<String> icp_string_split = this.toString().split('.');
+        if (icp_string_split.length <= 1) {
+            return this;
+        }
+        String decimal_places_string = icp_string_split[1];
+        if (round_decimal_places >= decimal_places_string.length) {
+            return this;
+        }
+        String decimal_places_split_at_round = decimal_places_string.substring(0, round_decimal_places);
+        return IcpTokens.oftheDoubleString('${icp_string_split[0]}.${BigInt.parse(decimal_places_split_at_round)+BigInt.from(1)}');
     }
     static IcpTokens oftheRecord(CandidType icptokensrecord) {
         Nat64 e8s_nat64 = (icptokensrecord as Record)['e8s'] as Nat64; 
@@ -274,13 +315,29 @@ class IcpTokens extends Record {
             e8s: e8s_nat64.value
         );
     }
-    static IcpTokens oftheDouble(double icp) {
-        if (check_double_decimal_point_places(icp) > IcpTokens.DECIMAL_PLACES) {
-            throw Exception('max ${IcpTokens.DECIMAL_PLACES} decimal places for the icp');
+    static IcpTokens oftheDoubleString(String icp_string) {
+        icp_string = icp_string.trim();
+        if (icp_string == '') {
+            throw Exception('must be a number');
         }
-        return IcpTokens(
-            e8s: BigInt.parse((icp * IcpTokens.DIVIDABLE_BY.toDouble()).toString().split('.')[0])
-        );
+        List<String> icp_string_split = icp_string.split('.');
+        if (icp_string_split.length > 2) {
+            throw Exception('invalid number.');
+        } 
+        BigInt icp = BigInt.parse(icp_string_split[0]);
+        BigInt icp_e8s_less_than_1 = BigInt.from(0);        
+        if (icp_string_split.length == 2) {
+            String decimal_places = icp_string_split[1];     
+            if (decimal_places.length > IcpTokens.DECIMAL_PLACES) {
+                throw Exception('Max ${IcpTokens.DECIMAL_PLACES} decimal places for the IcpTokens');
+            }
+            while (decimal_places.length < IcpTokens.DECIMAL_PLACES) {
+                decimal_places = '${decimal_places}0';
+            }
+            icp_e8s_less_than_1 = BigInt.parse(decimal_places);
+        }
+        IcpTokens icptokens = IcpTokens(e8s: (icp * IcpTokens.DIVIDABLE_BY) + icp_e8s_less_than_1);
+        return icptokens;
     }
     static int DECIMAL_PLACES = 8;    
     static BigInt DIVIDABLE_BY = BigInt.from(pow(10, IcpTokens.DECIMAL_PLACES));