+++ lib/src/tools/tools.dart
@@ -125,8 +125,6 @@ String bytes_as_the_bitstring(Uint8List bytes) {
 
 
 
-int check_double_decimal_point_places(double d) => d.toString().substring(d.toString().indexOf('.') + 1).length;
-
 
 String get_typename_ofthe_toString(String str) => str.substring(13, str.length-1);